--- planetwars/client/app/actions/matches/index.ts
@@ -1,2 +1,0 @@
-export * from './hosting';
-export * from './matches';
