+++ planetwars/client/app/components/index.ts
@@ -1,6 +1,5 @@
 export { default as BotsPage } from './bots/BotsPage';
 export { default as HomePage } from './home/HomePage';
-export { default as HostPage } from './host/HostPage';
 export { default as InfoPage } from './info/InfoPage';
 export { default as JoinPage } from './join/JoinPage';
 export { default as MatchesPage } from './matches/MatchesPage';
