+++ planetwars/client/app/reducers/root.ts
@@ -20,7 +20,6 @@ import { MatchesState, matchesReducer } from './matches';
 import { NotificationsState, notificationReducer } from './notifications';
 import { MapsState, mapsReducer } from './maps';
 
-import { HostPageState, hostReducer } from './hostPage';
 import { NavbarState, navbarReducer } from './navbar';
 import { LobbyState, lobbyReducer, defaultLobbyState } from './lobby';
 
@@ -33,7 +32,6 @@ export interface GState {
   readonly maps: MapsState;
   readonly lobby: LobbyState;
 
-  readonly host: HostPageState;
   readonly navbar: NavbarState;
 }
 
@@ -46,10 +44,6 @@ export const initialState: GState = {
   maps: {},
   lobby: defaultLobbyState,
   notifications: [],
-  host: {
-    slots: [],
-    serverRunning: false,
-  },
 };
 
 export const rootReducer = combineReducers<GState>({
@@ -61,6 +55,5 @@ export const rootReducer = combineReducers<GState>({
   notifications: notificationReducer,
   lobby: lobbyReducer,
 
-  host: hostReducer,
   navbar: navbarReducer,
 });
